generator/requirement.txt
absl-py==0.2.2
alabaster==0.7.10
amqp==2.2.2
anaconda-client==1.6.3
anaconda-navigator==1.6.2
anaconda-project==0.6.0
appnope==0.1.0
appscript==1.0.1
APScheduler==3.4.0
argh==0.26.2
arrow==0.14.2
asn1crypto==0.23.0
astor==0.6.2
astroid==1.4.9
astropy==1.3.2
Babel==2.4.0
backports.shutil-get-terminal-size==1.0.0
bcrypt==3.1.3
beautifulsoup4==4.6.0
billiard==3.5.0.3
bitarray==0.8.1
blaze==0.10.1
bleach==2.1
bokeh==0.12.5
boto==2.46.1
Bottleneck==1.2.1
breadability==0.1.20
bs4==0.0.1
bunch==1.0.1
bz2file==0.98
celery==4.1.0
certifi==2017.11.5
cffi==1.11.0
chardet==3.0.4
click==6.7
cloudpickle==0.2.2
clyent==1.2.2
colorama==0.3.9
conda==4.5.4
configparser==3.5.0
contextlib2==0.5.5
crayons==0.1.2
cryptography==2.0.3
cssselect==1.0.1
cycler==0.10.0
Cython==0.25.2
cytoolz==0.8.2
dash==0.19.0
dash-core-components==0.15.2
dash-html-components==0.8.0
dash-renderer==0.11.1
dask==0.14.3
datashape==0.5.4
decorator==4.1.2
defusedxml==0.5.0
distributed==1.16.3
docopt==0.6.2
docutils==0.13.1
entrypoints==0.2.3
et-xmlfile==1.0.1
Fabric3==1.13.1.post1
facebook-wda==0.2.2.dev1
fastcache==1.0.2
feedparser==5.2.1
Flask==1.0.3
Flask-Compress==1.4.0
Flask-Cors==3.0.2
Flask-Login==0.4.0
gast==0.2.0
gensim==2.3.0
gevent==1.2.1
greenlet==0.4.12
grpcio==1.12.1
h5py==2.7.0
HeapDict==1.0.0
html5lib==1.0b8
http-prompt==0.10.2
httpie==0.9.9
hugo-jupyter==1.0.2
idna==2.6
imagesize==0.7.1
ipykernel==4.6.1
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.0.1
isort==4.2.5
itsdangerous==0.24
jdcal==1.3
jedi==0.11.0
jieba==0.39
Jinja2==2.10.1
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.1.0
jupyter-console==5.2.0
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.3.1
jupyter-core==4.3.0
jupyter-highlight-selected-word==0.0.11
jupyter-latex-envs==1.3.8.4
jupyter-nbextensions-configurator==0.2.7
kaggle==1.3.5
Keras==2.1.6
kombu==4.1.0
lazy-object-proxy==1.2.2
llvmlite==0.18.0
locket==0.2.0
lxml==3.7.3
Markdown==2.6.11
MarkupSafe==1.0
marshmallow==3.0.0rc8
matplotlib==2.0.2
mistune==0.7.4
mmh3==2.5.1
mpmath==0.19
msgpack==0.5.6
msgpack-python==0.4.8
multipledispatch==0.4.9
mysql-connector-python-rf==2.2.2
mysqlclient==1.3.12
navigator-updater==0.1.0
nbconvert==5.3.1
nbformat==4.4.0
neovim==0.2.3
networkx==1.11
nltk==3.2.3
nose==1.3.7
notebook==5.1.0
numba==0.33.0
numexpr==2.6.2
numpy==1.14.5
numpydoc==0.6.0
odo==0.5.0
olefile==0.44
opencv-python==3.4.1.15
openpyxl==2.4.7
packaging==16.8
pandas==0.20.1
pandocfilters==1.4.2
paramiko==2.3.1
parsimonious==0.8.0
parso==0.1.0
partd==0.3.8
pathlib2==2.2.1
pathtools==0.1.2
patsy==0.4.1
pep8==1.7.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.1.1
pipreqs==0.4.9
plotly==2.2.3
ply==3.10
prompt-toolkit==1.0.15
protobuf==3.6.0
psutil==5.2.2
ptyprocess==0.5.2
py==1.4.33
pyasn1==0.3.6
pycosat==0.6.3
pycparser==2.18
pycrypto==2.6.1
pycurl==7.43.0
pydot==1.2.4
pyflakes==1.5.0
pygame==1.9.6
Pygments==2.2.0
pylint==1.6.4
pymongo==3.4.0
PyNaCl==1.1.2
pyodbc==4.0.16
pyOpenSSL==17.0.0
pyparsing==2.1.4
pyquery==1.2.17
pyspider==0.3.9
pytest==3.0.7
python-dateutil==2.6.1
python-votesmart==0.3.3
pytz==2018.3
PyWavelets==0.5.2
PyYAML==3.12
pyzmq==16.0.2
QtAwesome==0.4.4
qtconsole==4.3.1
QtPy==1.2.1
requests==2.18.4
rope-py3k==0.9.4.post1
scikit-image==0.13.0
scikit-learn==0.19.0
scipy==0.19.0
seaborn==0.7.1
setproctitle==1.1.10
simplegeneric==0.8.1
simplejson==3.13.2
singledispatch==3.4.0.3
six==1.11.0
smart-open==1.5.3
snowballstemmer==1.2.1
sortedcollections==0.5.3
sortedcontainers==1.5.7
Sphinx==1.5.6
spyder==3.1.4
SQLAlchemy==1.1.9
statsmodels==0.8.0
sumy==0.7.0
sympy==1.0
tables==3.3.0
tblib==1.3.2
tensorboard==1.8.0
tensorflow==1.8.0
termcolor==1.1.0
terminado==0.6
testpath==0.3.1
text-to-image==0.0.5
tinify==1.5.1
toolz==0.8.2
tornado==4.5.2
traitlets==4.3.2
tzlocal==1.4
u-msgpack-python==2.4.1
unicodecsv==0.14.1
urllib3==1.22
vine==1.1.4
watchdog==0.8.3
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.0.3
wrapt==1.10.10
WsgiDAV==2.2.2
xlrd==1.0.0
XlsxWriter==0.9.6
xlwings==0.10.4
xlwt==1.2.0
yarg==0.1.9
zict==0.1.2
